[integration-tests] try-fix integration tests
