+++ lib/modules/splash/view/splash_view.dart
@@ -1,11 +1,17 @@
+import 'package:encointer_wallet/mocks/substrate_api/core/mock_dart_api.dart';
+import 'package:encointer_wallet/mocks/substrate_api/mock_api.dart';
+import 'package:encointer_wallet/mocks/substrate_api/mock_js_api.dart';
+import 'package:encointer_wallet/page-encointer/home_page.dart';
+import 'package:encointer_wallet/page/account/create_account_entry_page.dart';
+import 'package:encointer_wallet/service/log/log_service.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
+import 'package:encointer_wallet/service/substrate_api/core/dart_api.dart';
+import 'package:encointer_wallet/service/substrate_api/core/js_api.dart';
+import 'package:encointer_wallet/store/app.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:provider/provider.dart';
-
-import 'package:encointer_wallet/page-encointer/home_page.dart';
-import 'package:encointer_wallet/page/account/create_account_entry_page.dart';
-import 'package:encointer_wallet/store/app.dart';
 
 class SplashView extends StatefulWidget {
   const SplashView({Key? key}) : super(key: key);
@@ -27,6 +33,15 @@ class _SplashViewState extends State<SplashView> {
     _store.dataUpdate.setupUpdateReaction(() async {
       await _store.encointer.updateState();
     });
+
+    final js = await DefaultAssetBundle.of(context).loadString('lib/js_service_encointer/dist/main.js');
+    webApi = (context.read<AppStore>().config == StoreConfig.Test
+        ? MockApi(context.read<AppStore>(), MockJSApi(), MockSubstrateDartApi(), js, withUi: true)
+        : Api.create(context.read<AppStore>(), JSApi(), SubstrateDartApi(), js))
+      ..init().timeout(
+        const Duration(seconds: 20),
+        onTimeout: () => Log.d('webApi.init() has run into a timeout. We might be offline.'),
+      );
 
     _store.setApiReady(true);
 
@@ -66,4 +81,16 @@ class _SplashViewState extends State<SplashView> {
       ),
     );
   }
+}
+
+Future<Api> initWebApi(BuildContext context) async {
+  final js = await DefaultAssetBundle.of(context).loadString('lib/js_service_encointer/dist/main.js');
+
+  return (context.read<AppStore>().config == StoreConfig.Test
+      ? MockApi(context.read<AppStore>(), MockJSApi(), MockSubstrateDartApi(), js, withUi: true)
+      : Api.create(context.read<AppStore>(), JSApi(), SubstrateDartApi(), js))
+    ..init().timeout(
+      const Duration(seconds: 20),
+      onTimeout: () => Log.d('webApi.init() has run into a timeout. We might be offline.'),
+    );
 }
